add test for alpha_s_bundle_pdf
